+++ src/yabc/basis.py
@@ -24,111 +24,8 @@ import yabc
 from yabc import csv_to_json
 from yabc import transaction
 from yabc.transaction import PreciseDecimalString
+from yabc.costbasisreport import CostBasisReport
 
-
-class CostBasisReport(yabc.Base):
-    """
-    Represents a row in form 8949.
-    TODO: Add wash sales.
-    TODO: Check that long term results match expected in leap years.
-
-    >>> r = CostBasisReport.make_random_report()
-    >>> r.date_sold = datetime.datetime(2019,1,1)
-    >>> r.date_purchased = datetime.datetime(2018,1,1)
-    >>> r._is_long_term()
-    True
-    >>> r.date_purchased = datetime.datetime(2018,1,2)
-    >>> r._is_long_term()
-    False
-    """
-
-    _fields = [
-        "id",
-        "asset_name",
-        "basis",
-        "date_purchased",
-        "date_sold",
-        "gain_or_loss",
-        "proceeds",
-        "quantity",
-        "long_term",
-        "adjustment",
-        "user_id",
-    ]
-
-    __tablename__ = "basis_report"
-    id = Column(Integer, primary_key=True)
-    asset_name = Column(sqlalchemy.String)
-    basis = Column(PreciseDecimalString)
-    date_purchased = Column(DateTime)
-    date_sold = Column(DateTime)
-    gain_or_loss = Column(PreciseDecimalString)
-    proceeds = Column(PreciseDecimalString)
-    quantity = Column(PreciseDecimalString)
-    adjustment = Column(PreciseDecimalString)
-    long_term = Column(Boolean)
-    user_id = Column(sqlalchemy.Integer, ForeignKey("user.id"))
-
-    @staticmethod
-    def make_random_report():
-        return CostBasisReport(
-            1,
-            Decimal(1),
-            Decimal(1),
-            datetime.datetime.now(),
-            Decimal(2),
-            datetime.datetime.now(),
-            "BTC",
-        )
-
-    def __init__(
-        self,
-        userid,
-        basis,
-        quantity,
-        date_purchased,
-        proceeds,
-        date_sold,
-        asset,
-        adjustment=Decimal(0),
-    ):
-        """
-        Note that when pulling items from a SQL alchemy ORM query, this constructor isn't called.
-        """
-        assert isinstance(date_sold, datetime.datetime)
-        assert isinstance(date_purchased, datetime.datetime)
-        assert isinstance(asset, str)
-        self.user_id = userid
-        self.basis = basis
-        self.quantity = quantity
-        self.date_purchased = date_purchased
-        self.proceeds = proceeds
-        self.date_sold = date_sold
-        self.asset_name = asset
-        self.gain_or_loss = self.proceeds - self.basis
-        self.adjustment = adjustment
-        self.long_term = self._is_long_term()
-
-    def _is_long_term(self):
-        return (self.date_sold - self.date_purchased) > datetime.timedelta(364)
-
-    def description(self):
-        return "{:.6f} {}".format(self.quantity, self.asset_name)
-
-    def __repr__(self):
-        return "<Sold {} {} for {} total profiting {}>".format(
-            self.quantity, self.asset_name, self.proceeds, self.gain_or_loss
-        )
-
-    def fields(self):
-        ans = []
-        for i in CostBasisReport._fields:
-            ans.append(getattr(self, i))
-        return ans
-
-    @staticmethod
-    def field_names():
-        return CostBasisReport._fields
 
 
 def split_coin_to_add(coin_to_split, amount, trans):
